// greenova/core/static/core/as/assembly/proto/landing/NewsletterSignupResponse.ts
// Code generated by protoc-gen-as. DO NOT EDIT.
// Versions:
//   protoc-gen-as v1.3.0
//   protoc        v4.24.4

import { Writer, Reader } from "as-proto/assembly";

export class NewsletterSignupResponse {
  static encode(message: NewsletterSignupResponse, writer: Writer): void {
    writer.uint32(8);
    writer.bool(message.success);

    writer.uint32(18);
    writer.string(message.message);
  }

  static decode(reader: Reader, length: i32): NewsletterSignupResponse {
    const end: usize = length < 0 ? reader.end : reader.ptr + length;
    const message = new NewsletterSignupResponse();

    while (reader.ptr < end) {
      const tag = reader.uint32();
      switch (tag >>> 3) {
        case 1:
          message.success = reader.bool();
          break;

        case 2:
          message.message = reader.string();
          break;

        default:
          reader.skipType(tag & 7);
          break;
      }
    }

    return message;
  }

  success: bool;
  message: string;

  constructor(success: bool = false, message: string = "") {
    this.success = success;
    this.message = message;
  }
}