// greenova/core/static/core/as/assembly/proto/landing/Stat.ts
// Code generated by protoc-gen-as. DO NOT EDIT.
// Versions:
//   protoc-gen-as v1.3.0
//   protoc        v4.24.4

import { Writer, Reader } from "as-proto/assembly";

export class Stat {
  static encode(message: Stat, writer: Writer): void {
    writer.uint32(10);
    writer.string(message.name);

    writer.uint32(16);
    writer.int32(message.value);
  }

  static decode(reader: Reader, length: i32): Stat {
    const end: usize = length < 0 ? reader.end : reader.ptr + length;
    const message = new Stat();

    while (reader.ptr < end) {
      const tag = reader.uint32();
      switch (tag >>> 3) {
        case 1:
          message.name = reader.string();
          break;

        case 2:
          message.value = reader.int32();
          break;

        default:
          reader.skipType(tag & 7);
          break;
      }
    }

    return message;
  }

  name: string;
  value: i32;

  constructor(name: string = "", value: i32 = 0) {
    this.name = name;
    this.value = value;
  }
}