// greenova/core/static/core/as/assembly/proto/landing/Testimonial.ts
// Code generated by protoc-gen-as. DO NOT EDIT.
// Versions:
//   protoc-gen-as v1.3.0
//   protoc        v4.24.4

import { Writer, Reader } from "as-proto/assembly";

export class Testimonial {
  static encode(message: Testimonial, writer: Writer): void {
    writer.uint32(10);
    writer.string(message.name);

    writer.uint32(18);
    writer.string(message.content);
  }

  static decode(reader: Reader, length: i32): Testimonial {
    const end: usize = length < 0 ? reader.end : reader.ptr + length;
    const message = new Testimonial();

    while (reader.ptr < end) {
      const tag = reader.uint32();
      switch (tag >>> 3) {
        case 1:
          message.name = reader.string();
          break;

        case 2:
          message.content = reader.string();
          break;

        default:
          reader.skipType(tag & 7);
          break;
      }
    }

    return message;
  }

  name: string;
  content: string;

  constructor(name: string = "", content: string = "") {
    this.name = name;
    this.content = content;
  }
}