// greenova/core/static/core/as/assembly/proto/landing/LandingPageContent.ts
// Code generated by protoc-gen-as. DO NOT EDIT.
// Versions:
//   protoc-gen-as v1.3.0
//   protoc        v4.24.4

import { Writer, Reader } from "as-proto/assembly";
import { Feature } from "./Feature";
import { Stat } from "./Stat";
import { Testimonial } from "./Testimonial";

export class LandingPageContent {
  static encode(message: LandingPageContent, writer: Writer): void {
    writer.uint32(10);
    writer.string(message.heroTitle);

    writer.uint32(18);
    writer.string(message.heroSubtitle);

    const features = message.features;
    for (let i: i32 = 0; i < features.length; ++i) {
      writer.uint32(26);
      writer.fork();
      Feature.encode(features[i], writer);
      writer.ldelim();
    }

    const stats = message.stats;
    for (let i: i32 = 0; i < stats.length; ++i) {
      writer.uint32(34);
      writer.fork();
      Stat.encode(stats[i], writer);
      writer.ldelim();
    }

    const benefits = message.benefits;
    if (benefits.length !== 0) {
      for (let i: i32 = 0; i < benefits.length; ++i) {
        writer.uint32(42);
        writer.string(benefits[i]);
      }
    }

    const testimonials = message.testimonials;
    for (let i: i32 = 0; i < testimonials.length; ++i) {
      writer.uint32(50);
      writer.fork();
      Testimonial.encode(testimonials[i], writer);
      writer.ldelim();
    }

    writer.uint32(58);
    writer.string(message.ctaTitle);

    writer.uint32(66);
    writer.string(message.ctaSubtitle);
  }

  static decode(reader: Reader, length: i32): LandingPageContent {
    const end: usize = length < 0 ? reader.end : reader.ptr + length;
    const message = new LandingPageContent();

    while (reader.ptr < end) {
      const tag = reader.uint32();
      switch (tag >>> 3) {
        case 1:
          message.heroTitle = reader.string();
          break;

        case 2:
          message.heroSubtitle = reader.string();
          break;

        case 3:
          message.features.push(Feature.decode(reader, reader.uint32()));
          break;

        case 4:
          message.stats.push(Stat.decode(reader, reader.uint32()));
          break;

        case 5:
          message.benefits.push(reader.string());
          break;

        case 6:
          message.testimonials.push(
            Testimonial.decode(reader, reader.uint32())
          );
          break;

        case 7:
          message.ctaTitle = reader.string();
          break;

        case 8:
          message.ctaSubtitle = reader.string();
          break;

        default:
          reader.skipType(tag & 7);
          break;
      }
    }

    return message;
  }

  heroTitle: string;
  heroSubtitle: string;
  features: Array<Feature>;
  stats: Array<Stat>;
  benefits: Array<string>;
  testimonials: Array<Testimonial>;
  ctaTitle: string;
  ctaSubtitle: string;

  constructor(
    heroTitle: string = "",
    heroSubtitle: string = "",
    features: Array<Feature> = [],
    stats: Array<Stat> = [],
    benefits: Array<string> = [],
    testimonials: Array<Testimonial> = [],
    ctaTitle: string = "",
    ctaSubtitle: string = ""
  ) {
    this.heroTitle = heroTitle;
    this.heroSubtitle = heroSubtitle;
    this.features = features;
    this.stats = stats;
    this.benefits = benefits;
    this.testimonials = testimonials;
    this.ctaTitle = ctaTitle;
    this.ctaSubtitle = ctaSubtitle;
  }
}